Fix updating of tier's end times.
End times of all tiers are now updated as part of TextGrid's __str__
method. Before this is done, copies of all tiers are made to avoid
altering end times of the original tiers.

Additionally, inserting empty intervals is now delegated to a separate
method of IntervalTier, which returns a copy of the original tier. As a
result, IntervalTier can now use the generic Tier __str__ method.